Refactored app-tools
